Add assertions for important assumptions
There previously was no assertions to
ensure that specific lines run appropriately.

This could make code hard to debug in the event
of an error.

Assertions are added in getResponse() and
checkInput() methods of Phi class, to check for:

getResponse():
- Input is not null

checkInput():
- Check that the size of the list of tasks change as expected

Such assertions would be helpful to ensure
that the code could be debugged properly should
it crash unexpectedly.